refactor(losses): isolate loss implementations and utilities
