+++ src/speculators/losses/__init__.py
@@ -1,0 +1,35 @@
+"""Training loss implementations and shared utilities."""
+
+from speculators.losses.utils import (
+    LossConfig,
+    ce_loss_fused_or_eager,
+    compound_loss,
+    dflash_loss_decay,
+    dpace_loss_decay,
+    exp_loss_decay,
+    js_div_loss_fused_or_eager,
+    kl_div_loss_fused_or_eager,
+    lk_hybrid_loss_fused_or_eager,
+    loss_function,
+    nla_loss_fused_or_eager,
+    resolve_loss_config,
+    reverse_kl_div_loss_fused_or_eager,
+    tv_loss_fused_or_eager,
+)
+
+__all__ = [
+    "LossConfig",
+    "ce_loss_fused_or_eager",
+    "compound_loss",
+    "dflash_loss_decay",
+    "dpace_loss_decay",
+    "exp_loss_decay",
+    "js_div_loss_fused_or_eager",
+    "kl_div_loss_fused_or_eager",
+    "lk_hybrid_loss_fused_or_eager",
+    "loss_function",
+    "nla_loss_fused_or_eager",
+    "resolve_loss_config",
+    "reverse_kl_div_loss_fused_or_eager",
+    "tv_loss_fused_or_eager",
+]
