+++ src/speculators/models/dspark/core.py
@@ -3,16 +3,16 @@ from typing import ClassVar
 import torch
 from transformers import PretrainedConfig
 
+from speculators.losses import (
+    LossConfig,
+    kl_div_loss_fused_or_eager,
+    resolve_loss_config,
+)
 from speculators.model import SpeculatorModel
 from speculators.models.dflash.core import DFlashDraftModel
 from speculators.models.dspark.config import DSparkSpeculatorConfig
 from speculators.models.dspark.metrics import compute_metrics
 from speculators.models.dspark.model_definitions import ConfidenceHead, MarkovHead
-from speculators.models.metrics import (
-    LossConfig,
-    kl_div_loss_fused_or_eager,
-    resolve_loss_config,
-)
 from speculators.models.utils import conditional_torch_compile
 
 _DEFAULT_LOSS_CONFIG: LossConfig = {"kl_div": (kl_div_loss_fused_or_eager, 1.0)}
