+++ src/speculators/models/metrics.py
@@ -1,15 +1,6 @@
-import json
-import math
-from collections.abc import Callable
-from functools import cache
+"""Shared model accuracy metrics."""
 
 import torch
-
-_EPS = 1e-5
-
-LossConfig = dict[
-    str, tuple[Callable[[torch.Tensor, torch.Tensor], torch.Tensor], float]
-]
 
 
 def compute_accuracy_single_step(
@@ -76,480 +67,3 @@ def compute_accuracy_multi_step(
     total_per_pos.scatter_add_(0, pos_idx, torch.ones_like(correct, dtype=torch.float))
 
     return correct_per_pos, total_per_pos  # shape: [num_pos], [num_pos]
-
-
-def kl_div_loss(
-    logits: torch.Tensor,  # shape: [1, seq_len, draft_vocab_size]
-    targets: torch.Tensor,  # shape: [1, seq_len, draft_vocab_size]
-):
-    """Compute per-position KL divergence from draft logits to target logits.
-
-    Args:
-        logits: Draft model logits (log-softmax applied internally).
-        targets: Target model logits (softmax applied internally).
-
-    Returns:
-        Per-position KL divergence with shape [1, seq_len].
-    """
-    logits = torch.nn.functional.log_softmax(logits, dim=-1, dtype=torch.float32)
-    target_p = torch.nn.functional.softmax(targets, dim=-1, dtype=torch.float32)
-    elementwise_loss = torch.nn.functional.kl_div(
-        logits, target_p, reduction="none", log_target=False
-    ).sum(dim=-1)  # shape: [1, seq_len]
-
-    return elementwise_loss  # noqa: RET504
-
-
-def reverse_kl_div_loss(
-    logits: torch.Tensor,  # shape: [1, seq_len, draft_vocab_size]
-    targets: torch.Tensor,  # shape: [1, seq_len, draft_vocab_size]
-):
-    """Compute per-position reverse KL divergence from draft logits to target logits.
-
-    Args:
-        logits: Draft model logits (log-softmax applied internally).
-        targets: Target model logits (log-softmax applied internally).
-
-    Returns:
-        Per-position reverse KL divergence with shape [1, seq_len].
-    """
-    draft_logq = torch.nn.functional.log_softmax(logits, dim=-1, dtype=torch.float32)
-    target_logp = torch.nn.functional.log_softmax(targets, dim=-1, dtype=torch.float32)
-    elementwise_loss = torch.nn.functional.kl_div(
-        target_logp, draft_logq, reduction="none", log_target=True
-    ).sum(dim=-1)  # shape: [1, seq_len]
-
-    return elementwise_loss  # noqa: RET504
-
-
-def js_div_loss(
-    logits: torch.Tensor,  # shape: [1, seq_len, draft_vocab_size]
-    targets: torch.Tensor,  # shape: [1, seq_len, draft_vocab_size]
-):
-    """Compute per-position Jensen-Shannon divergence between draft and target.
-
-    ``JSD(p, q) = 0.5 * KL(p || m) + 0.5 * KL(q || m)`` with ``m = (p + q) / 2``.
-    Symmetric and bounded by ``log 2`` (Lin 1991, "Divergence measures based on
-    the Shannon entropy"), it balances forward KL's mass-covering pull with
-    reverse KL's mode-seeking pull and keeps gradients finite where either
-    distribution assigns near-zero probability. Compared to plain KL, this
-    avoids unbounded penalties on tokens the target barely supports; compared
-    to TV, it provides smoother, better-conditioned gradients for draft
-    training.
-
-    Args:
-        logits: Draft model logits (log-softmax applied internally).
-        targets: Target model logits (log-softmax applied internally).
-
-    Returns:
-        Per-position JS divergence with shape [1, seq_len].
-    """
-    draft_logq = torch.nn.functional.log_softmax(logits, dim=-1, dtype=torch.float32)
-    target_logp = torch.nn.functional.log_softmax(targets, dim=-1, dtype=torch.float32)
-    # log m = log((p + q) / 2), computed in log space for stability
-    log_m = torch.logaddexp(draft_logq, target_logp) - math.log(2.0)
-    kl_target_to_mix = torch.nn.functional.kl_div(
-        log_m, target_logp, reduction="none", log_target=True
-    ).sum(dim=-1)
-    kl_draft_to_mix = torch.nn.functional.kl_div(
-        log_m, draft_logq, reduction="none", log_target=True
-    ).sum(dim=-1)
-    elementwise_loss = 0.5 * (kl_target_to_mix + kl_draft_to_mix)  # [1, seq_len]
-
-    return elementwise_loss  # noqa: RET504
-
-
-def ce_loss(
-    logits: torch.Tensor,  # shape: [1, seq_len, draft_vocab_size]
-    targets: torch.Tensor,  # shape: [1, seq_len, draft_vocab_size]
-):
-    """Compute per-position cross-entropy loss using argmax of target logits as labels.
-
-    Args:
-        logits: Draft model logits.
-        targets: Target model logits (argmax taken to produce hard labels).
-
-    Returns:
-        Per-position cross-entropy loss with shape [1, seq_len].
-    """
-    batch_size, seq_len, draft_vocab_size = logits.shape
-    target_ids = torch.argmax(targets, dim=-1)  # shape: [1, seq_len]
-
-    elementwise_loss = torch.nn.functional.cross_entropy(
-        logits.reshape(-1, draft_vocab_size),
-        target_ids.reshape(-1),
-        reduction="none",
-        ignore_index=-100,
-    ).reshape(batch_size, seq_len)
-
-    return elementwise_loss  # noqa: RET504
-
-
-def tv_loss(
-    logits: torch.Tensor,  # shape: [1, seq_len, draft_vocab_size]
-    targets: torch.Tensor,  # shape: [1, seq_len, draft_vocab_size]
-):
-    """Compute per-position total variation (TV) distance from draft to target.
-
-    The rejection-sampling acceptance rate of speculative decoding equals the
-    distributional overlap between target and draft,
-    ``alpha = sum_v min(p_v, q_v) = 1 - d_TV(p, q)``. Minimizing this TV distance
-    therefore directly optimizes the acceptance rate, whereas cross-entropy and
-    KL only optimize it indirectly (KL is a loose upper bound on TV via Pinsker).
-
-    Args:
-        logits: Draft model logits (softmax applied internally to form q).
-        targets: Target model logits (softmax applied internally to form p).
-
-    Returns:
-        Per-position TV distance with shape [1, seq_len].
-    """
-    draft_p = torch.nn.functional.softmax(logits, dim=-1, dtype=torch.float32)
-    target_p = torch.nn.functional.softmax(targets, dim=-1, dtype=torch.float32)
-    overlap = torch.minimum(draft_p, target_p).sum(dim=-1)  # shape: [1, seq_len]
-    elementwise_loss = 1.0 - overlap
-
-    return elementwise_loss  # noqa: RET504
-
-
-def neg_log_acceptance_loss(
-    logits: torch.Tensor,  # shape: [1, seq_len, draft_vocab_size]
-    targets: torch.Tensor,  # shape: [1, seq_len, draft_vocab_size]
-):
-    """Compute per-position negative log-acceptance (LK) loss.
-
-    The speculative-decoding acceptance rate equals the draft/target distribution
-    overlap, ``alpha = sum_v min(p_v, q_v)`` (the same quantity computed in
-    ``tv_loss``). This loss is ``-log(alpha)``. Its gradient is
-    ``(1 / alpha) * grad(TV)``: the ``1 / alpha`` factor amplifies the otherwise
-    vanishing TV gradient when overlap is low (early training), giving TV's
-    acceptance-optimal target a usable gradient from a cold start. When the target
-    is a point mass, this loss reduces to cross-entropy.
-
-    Args:
-        logits: Draft model logits (softmax applied internally to form q).
-        targets: Target model logits (softmax applied internally to form p).
-
-    Returns:
-        Per-position negative log-acceptance with shape [1, seq_len].
-    """
-    draft_p = torch.nn.functional.softmax(logits, dim=-1, dtype=torch.float32)
-    target_p = torch.nn.functional.softmax(targets, dim=-1, dtype=torch.float32)
-    overlap = torch.minimum(draft_p, target_p).sum(dim=-1)  # alpha, shape: [1, seq_len]
-    elementwise_loss = -torch.log(overlap.clamp_min(_EPS))
-
-    return elementwise_loss  # noqa: RET504
-
-
-def lk_hybrid_loss(
-    logits: torch.Tensor,  # shape: [1, seq_len, draft_vocab_size]
-    targets: torch.Tensor,  # shape: [1, seq_len, draft_vocab_size]
-    eta: float = 3.0,
-):
-    """Compute per-position hybrid LK loss (adaptive KL/TV blend).
-
-    Blends KL divergence and total variation per position:
-    ``L = lambda * KL(p||q) + (1 - lambda) * TV(p, q)`` with adaptive weight
-    ``lambda = exp(-eta * sg[alpha])``, where ``alpha = sum_v min(p_v, q_v)`` is the
-    acceptance rate (overlap) and ``sg`` is stop-gradient. When overlap is low
-    (early training, misaligned draft) ``lambda -> 1`` and the loss leans on KL's
-    strong gradient; as overlap grows ``lambda -> 0`` and it shifts to TV, which
-    optimizes acceptance directly. This gives TV's acceptance-optimal target a
-    usable gradient from a cold start.
-
-    ``alpha`` in the weight is detached: it controls the blend but is not
-    differentiated through; gradients flow only through the KL and TV terms.
-
-    Source: Samarin et al., "LK Losses: Direct Acceptance Rate Optimization for
-    Speculative Decoding" (arXiv 2602.23881), hybrid objective.
-
-    Args:
-        logits: Draft model logits (softmax applied internally to form q).
-        targets: Target model logits (softmax applied internally to form p).
-        eta: Blend temperature; larger shifts toward TV sooner. Default 3.0
-            (the paper's best hybrid setting).
-
-    Returns:
-        Per-position hybrid loss with shape [1, seq_len].
-    """
-    draft_p = torch.nn.functional.softmax(logits, dim=-1, dtype=torch.float32)
-    target_p = torch.nn.functional.softmax(targets, dim=-1, dtype=torch.float32)
-    overlap = torch.minimum(draft_p, target_p).sum(dim=-1)  # alpha, shape: [1, seq_len]
-    tv = 1.0 - overlap
-    kl = kl_div_loss(logits, targets)  # reuse existing KL, shape: [1, seq_len]
-    weight = torch.exp(-eta * overlap.detach())  # lambda = exp(-eta * sg[alpha])
-    elementwise_loss = weight * kl + (1.0 - weight) * tv
-
-    return elementwise_loss  # noqa: RET504
-
-
-def dflash_loss_decay(
-    pos_idx: torch.Tensor, gamma: float, sample_from_anchor: bool = False, **_kwargs
-):
-    """Compute DFlash-style exponential decay weights per position.
-
-    When ``sample_from_anchor`` is False (DFlash), position 0 is the anchor and
-    gets weight 0; position 1 gets weight 1, and subsequent positions decay as
-    ``exp(-(pos - 1) / gamma)``.
-
-    When ``sample_from_anchor`` is True (DSpark), position 0 is the first real
-    prediction and gets weight 1; subsequent positions decay as
-    ``exp(-pos / gamma)``.
-
-    Args:
-        pos_idx: Position indices within each speculative block.
-        gamma: Decay rate (higher = slower decay).
-        sample_from_anchor: Whether position 0 is a prediction (DSpark) or the
-            anchor input (DFlash).
-
-    Returns:
-        Decay multiplier tensor with same shape as pos_idx.
-    """
-    if sample_from_anchor:
-        # DSpark: slot 0 predicts the token after the anchor, so decay from 0.
-        return torch.exp(-pos_idx / gamma)
-    # DFlash: slot 0 is the anchor (no prediction); decay from slot 1.
-    # pos_idx = 0 1 2 3 0 1 2 3, block_size = 4
-    decay_mult = torch.exp(-((pos_idx - 1).clamp(min=0)) / gamma)
-    # decay_mult = e^-(0 0 1 2 0 0 1 2) / gamma
-    decay_mult = decay_mult * (pos_idx != 0).to(decay_mult.dtype)
-    # w = 0 1 e^-1/gamma e^-2/gamma 0 1 e^-1/gamma e^-2/gamma
-    return decay_mult  # noqa: RET504
-
-
-def exp_loss_decay(pos_idx: torch.Tensor, gamma: float, **_kwargs):
-    """Compute simple exponential decay weights as gamma^pos_idx.
-
-    Args:
-        pos_idx: Position indices within each speculative block.
-        gamma: Base of the exponent (typically in (0, 1]).
-
-    Returns:
-        Decay multiplier tensor with same shape as pos_idx.
-    """
-    return gamma**pos_idx
-
-
-def dpace_loss_decay(
-    pos_idx: torch.Tensor,  # noqa: ARG001
-    loss_mask: torch.Tensor,
-    block_size: int,
-    dpace_alpha: float,
-    elementwise_loss: torch.Tensor,
-    **_kwargs,
-):
-    """
-    Per-position block-drafting loss weight based on D-PACE
-
-    Args:
-        elementwise_loss: requires to be cross-entropy loss, negative log-likelihood
-            of per-position confidence
-        dpace_alpha: confidence smoothing constant
-
-    Returns:
-        Decay multiplier tensor with same shape as pos_idx.
-    """
-    with torch.no_grad():
-        # convert CE to per-position confidence
-        q = torch.exp(-elementwise_loss).float()
-
-        # reshape loss to [num_anchors, block_size]
-        # for intra-block cumulative multiplication
-        if q.shape[1] % block_size != 0:
-            raise ValueError(
-                f"q.shape[1] ({q.shape[1]}) must be divisible by "
-                f"block_size ({block_size})"
-            )
-        num_anchors = q.shape[1] // block_size
-        q = q.reshape(num_anchors, block_size)
-        mask = loss_mask.reshape(num_anchors, block_size).to(q.dtype)
-
-        # smoothed confidence for numerical stability
-        smooth = (1.0 - dpace_alpha) * q + dpace_alpha
-        smooth = torch.where(mask > 0, smooth, torch.ones_like(smooth))
-
-        # prefix cumulative production
-        prefix = torch.cumprod(smooth, dim=-1)
-
-        # suffix summation: flip -> cumsum -> flip
-        weight = torch.flip(
-            torch.cumsum(torch.flip(prefix * mask, dims=[-1]), dim=-1), dims=[-1]
-        )
-        weight = weight * mask
-
-    # reshape weight
-    return weight.reshape(1, -1)
-
-
-# Fused Triton losses run on CUDA/ROCm (a few scalars per row instead of fp32
-# [T, V] intermediates; see models/fused_losses.py); everything else -- CPU,
-# NPUs without a Triton backend -- falls back to the eager losses above. The
-# import is lazy so this module works without Triton installed.
-
-
-@cache
-def _fused_kernel(name: str):
-    """Import and cache a fused loss by name; ``None`` if Triton is unavailable."""
-    try:
-        from speculators.models import fused_losses as mod  # noqa: PLC0415
-    except ImportError:
-        return None
-    return getattr(mod, name)
-
-
-def _fused_or_eager(
-    fused_name: str,
-    eager_fn: Callable[[torch.Tensor, torch.Tensor], torch.Tensor],
-) -> Callable[[torch.Tensor, torch.Tensor], torch.Tensor]:
-    """Dispatch to the fused Triton twin on CUDA/ROCm, else run *eager_fn*.
-
-    Fused backward returns no target gradient, so differentiable targets also
-    fall back to eager (training targets come from a no-grad verifier).
-    """
-
-    def loss_fn(logits: torch.Tensor, targets: torch.Tensor) -> torch.Tensor:
-        if logits.is_cuda and not targets.requires_grad:
-            fused = _fused_kernel(fused_name)
-            if fused is not None:
-                return fused(logits, targets)
-        return eager_fn(logits, targets)
-
-    loss_fn.__name__ = f"{eager_fn.__name__}_fused_or_eager"
-    return loss_fn
-
-
-kl_div_loss_fused_or_eager = _fused_or_eager("fused_kl_div_loss", kl_div_loss)
-reverse_kl_div_loss_fused_or_eager = _fused_or_eager(
-    "fused_reverse_kl_div_loss", reverse_kl_div_loss
-)
-js_div_loss_fused_or_eager = _fused_or_eager("fused_js_div_loss", js_div_loss)
-ce_loss_fused_or_eager = _fused_or_eager("fused_ce_loss", ce_loss)
-tv_loss_fused_or_eager = _fused_or_eager("fused_tv_loss", tv_loss)
-nla_loss_fused_or_eager = _fused_or_eager("fused_nla_loss", neg_log_acceptance_loss)
-lk_hybrid_loss_fused_or_eager = _fused_or_eager("fused_lk_hybrid_loss", lk_hybrid_loss)
-
-
-_LOSS_FN_MAP: dict[str, Callable[[torch.Tensor, torch.Tensor], torch.Tensor]] = {
-    "kl_div": kl_div_loss_fused_or_eager,
-    "rkl": reverse_kl_div_loss_fused_or_eager,
-    "jsd": js_div_loss_fused_or_eager,
-    "ce": ce_loss_fused_or_eager,
-    "tv": tv_loss_fused_or_eager,
-    "nla": nla_loss_fused_or_eager,
-    "lk_hybrid": lk_hybrid_loss_fused_or_eager,
-}
-
-
-def resolve_loss_config(spec: str) -> LossConfig:
-    """Parse a loss spec into ``{name: (loss_fn, weight)}``.
-
-    Accepts either a plain loss name (``"kl_div"``) or a JSON dict mapping
-    loss names to weights (``'{"ce": 0.1, "tv": 0.9}'``).
-    """
-    if spec in _LOSS_FN_MAP:
-        return {spec: (_LOSS_FN_MAP[spec], 1.0)}
-
-    try:
-        parsed = json.loads(spec)
-    except json.JSONDecodeError:
-        raise ValueError(
-            f"Unknown loss function '{spec}'. Pass a known name "
-            f"({sorted(_LOSS_FN_MAP.keys())}) or a JSON dict, "
-            f'e.g. \'{{"ce": 0.1, "tv": 0.9}}\'.'
-        ) from None
-
-    if not isinstance(parsed, dict) or not parsed:
-        raise ValueError(
-            "Loss config must be a non-empty JSON dict mapping loss names to weights, "
-            f'e.g. \'{{"ce": 0.1, "tv": 0.9}}\'. Got: {spec}'
-        )
-
-    config: LossConfig = {}
-    for name, weight in parsed.items():
-        if name not in _LOSS_FN_MAP:
-            raise ValueError(
-                f"Unknown loss function '{name}' in loss config. "
-                f"Choose from: {sorted(_LOSS_FN_MAP.keys())}"
-            )
-        if not isinstance(weight, (int, float)):
-            raise ValueError(
-                f"Loss weight for '{name}' must be a number, "
-                f"got {type(weight).__name__}"
-            )
-        config[name] = (_LOSS_FN_MAP[name], float(weight))
-
-    return config
-
-
-def compound_loss(
-    logits: torch.Tensor,
-    targets: torch.Tensor,
-    loss_mask: torch.Tensor,
-    pos_idx: torch.Tensor,
-    loss_config: LossConfig,
-    decay_fn: Callable[..., torch.Tensor] | None = None,
-) -> tuple[torch.Tensor, dict[str, torch.Tensor]]:
-    """Compute a weighted sum of loss terms.
-
-    Each entry in *loss_config* maps a name to ``(loss_fn, weight)``; the
-    result is ``sum(weight * loss_function(logits, targets, ..., loss_fn))``
-    over all entries.
-
-    Returns the total loss and a dict of per-term (unweighted) scalar losses
-    keyed as ``"{name}_loss"``.  When the config contains a single term the
-    dict is empty (the overall loss already captures it).
-    """
-    total = torch.tensor(0.0, device=logits.device, dtype=torch.float32)
-    term_losses: dict[str, torch.Tensor] = {}
-    multi = len(loss_config) > 1
-    for name, (fn, weight) in loss_config.items():
-        term = loss_function(
-            logits,
-            targets,
-            loss_mask,
-            pos_idx,
-            loss_fn=fn,
-            decay_fn=decay_fn,
-        )
-        if multi:
-            term_losses[f"{name}_loss"] = term.detach()
-        total = total + weight * term
-    return total, term_losses
-
-
-def loss_function(
-    logits: torch.Tensor,  # shape: [1, seq_len, draft_vocab_size]
-    targets: torch.Tensor,  # shape: [1, seq_len, draft_vocab_size]
-    loss_mask: torch.Tensor,  # shape: [1, seq_len]
-    pos_idx: torch.Tensor,  # shape: [1, seq_len]
-    loss_fn: Callable[[torch.Tensor, torch.Tensor], torch.Tensor] = kl_div_loss,
-    decay_fn: Callable[..., torch.Tensor] | None = None,
-):
-    """Compute masked, optionally position-decayed training loss.
-
-    Args:
-        logits: Draft model logits.
-        targets: Target model logits.
-        loss_mask: Boolean mask selecting positions to include in the loss.
-        pos_idx: Position indices within each speculative block.
-        loss_fn: Per-position loss function (default: kl_div_loss).
-        decay_fn: Optional position-dependent decay weighting function.
-
-    Returns:
-        Scalar mean loss across the batch.
-    """
-    elementwise_loss = loss_fn(logits, targets)  # shape: [1, seq_len]
-
-    loss_mask = loss_mask.to(elementwise_loss.dtype)
-    elementwise_loss = elementwise_loss * loss_mask
-
-    if decay_fn is not None:
-        decay_mult = decay_fn(
-            pos_idx.to(elementwise_loss.dtype), elementwise_loss=elementwise_loss
-        )
-        elementwise_loss = elementwise_loss * decay_mult
-
-    denominator = loss_mask.sum(dim=1) + _EPS
-
-    batch_loss = torch.sum(elementwise_loss, dim=1) / denominator  # shape: [1]
-    return batch_loss.mean()  # shape: []
